% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Generalized Linear Regression with Little Bag of Bootstraps}
\usage{
blbglm(formula, data, family = "gaussian", m = 10, B = 5000)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{m}{The amounts of subsamples to create}

\item{B}{The times of boostrapy
#' @export}
}
\description{
This function implement the GLM model by blb method
}
