% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm_prallel}
\alias{blbglm_prallel}
\title{Generalized Linear Regression with Little Bag of Bootstraps parallely}
\usage{
blbglm_prallel(formula, data, family = "gaussian", m = 10, B = 5000, core = 1)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{m}{The amounts of subsamples to create}

\item{B}{The times of boostrapy}

\item{core}{The numbers of cores to use for parallel computation}
}
\description{
This function implement the GLM model by blb method with parallel computation
}
