% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The formula to fit.}

\item{data}{The data from which the regression is fitted}

\item{m}{The number of bags to divid to}

\item{B}{The amount of times boostrap will run}
}
\description{
The core function to implement linear Regression with blb.
}
