% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{blblm_parallel}
\alias{blblm_parallel}
\title{Linear Regression with Parallel Little Bag of Bootstraps by parallel package}
\usage{
blblm_parallel(formula, data, m = 10, B = 5000, core = 1)
}
\arguments{
\item{formula}{The formula to fit.}

\item{data}{The data from which the regression is fitted}

\item{m}{The number of bags to divid to}

\item{B}{The amount of times boostrap will run}

\item{core}{The number of core parallel computation will use}
}
\description{
The core function to parallely implement linear Regression with blb by parallel
}
