% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_robust.R
\name{blblm_robust}
\alias{blblm_robust}
\title{Robust Linear Regression with Little Bag of Bootstraps}
\usage{
blblm_robust(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The formula to fit.}

\item{data}{The data from which the regression is fitted}

\item{m}{The number of bags to divid to}

\item{B}{The amount of times boostrap will run}
}
\description{
The function to implement robust linear Regression fitting with blb.
}
