% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{blbsigma_rcpp}
\alias{blbsigma_rcpp}
\title{compute sigma from fit}
\usage{
blbsigma_rcpp(fit, weights, y)
}
\arguments{
\item{fit}{The fitted model}

\item{weights}{The weights used to fit}

\item{y}{The response varibale}
}
\description{
compute sigma from fit
}
