% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get the coefficiences of the model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{The mode you fitted}

\item{...}{not use}
}
\description{
Get the coefficiences of the model
}
