% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Get the confidence interval of the model}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The model you fitted}

\item{parm}{The specific parameter you want to use}

\item{level}{significant level}

\item{...}{not use}
}
\description{
Get the confidence interval of the model
}
