% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{The linear function to fit}

\item{data}{The data inputed}

\item{freqs}{The frequences to boot the data}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
