% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{The linear function to fit}

\item{data}{The data inputed}

\item{n}{The numbers of rows to be sampled from}
}
\description{
compute the regression estimates for a blb dataset
}
