% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{The linear function to fit}

\item{data}{The data inputed}

\item{n}{The numbers of rows to be sampled from}

\item{B}{The numbers of trails}
}
\description{
compute the estimates
}
