% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict through the fitted model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object, new_data, confidence, level}{data}

\item{...}{not use}
}
\description{
Predict through the fitted model
}
