% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Get the sigma of the model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The fitted model}

\item{confidence}{Whether you want CI or not}

\item{level}{significant level for computing CI}

\item{...}{not use}
}
\description{
Get the sigma of the model
}
