% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_core.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{The input data}

\item{m}{The numbers of subsamples to create}
}
\description{
split data into m parts of approximated equal sizes
}
